requirements.txt
av==14.0.1
certifi==2024.12.14
charset-normalizer==3.4.0
click==8.1.8
dlib==19.24.6
face-recognition==1.3.0
face-recognition-models== 0.3.0
facenet-pytorch==2.6.0
filelock==3.16.1
fsspec==2024.12.0
idna==3.10
Jinja2==3.1.5
joblib==1.4.2
MarkupSafe==3.0.2
mpmath==1.3.0
psutil==6.1.1
networkx==3.4.2
numpy==1.26.4
opencv-contrib-python==4.10.0.84
opencv-python==4.10.0.84
pillow==10.2.0
PyYAML==6.0.2
requests==2.32.3
beautifulsoup4==4.12.3
scikit-learn==1.6.0
scipy==1.14.1
setuptools==68.2.0
sympy==1.13.3
threadpoolctl==3.5.0
torch==2.2.2
torchaudio==2.2.2
torchvision==0.17.2
tqdm==4.67.1
typing_extensions==4.12.2
textual==0.55.1
urllib3==2.3.0
wheel==0.41.2
pytest==8.3.4
pytest-emoji==0.2.0
pytest-md==0.2.0
